Hide suggested todos until you have completed the initial todos.
