+++ front/components/assistant/conversation/space/conversations/project_todos/ProjectTodoScopeFilter.tsx
@@ -3,7 +3,6 @@ import {
   Avatar,
   Button,
   ChevronDownIcon,
-  cn,
   DropdownMenu,
   DropdownMenuCheckboxItem,
   DropdownMenuContent,
@@ -18,7 +17,6 @@ import {
 
 export function ProjectTodoScopeFilter() {
   const {
-    showSuggestedTodosTable,
     assigneeSearch,
     setAssigneeSearch,
     isAssigneeMenuOpen,
@@ -36,13 +34,7 @@ export function ProjectTodoScopeFilter() {
   } = useProjectTodosPanel();
 
   return (
-    <div
-      className={cn(
-        "border-b border-border pb-6 dark:border-border-night",
-        !showSuggestedTodosTable &&
-          "border-t border-border pt-6 dark:border-border-night"
-      )}
-    >
+    <div className="border-b border-border pb-6 dark:border-border-night">
       <div className="inline-flex w-full items-center gap-2">
         <DropdownMenu
           modal={false}
