+++ front/components/assistant/conversation/space/conversations/project_todos/useProjectTodosPanelState.ts
@@ -627,6 +627,7 @@ export function useProjectTodosPanelState({
     handleAddTodo,
     isTodosLoading,
     frozenLastReadAt,
+    todos,
     filteredTodos,
   };
 }
